requirements.txt
Flask
flask-cors
python-dotenv
google-generativeai
faiss-cpu
numpy
